apps/api/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.5

# Settings
pydantic-settings==2.5.2
pydantic==2.9.2
python-dotenv==1.0.1

# LLM / Embedding (OpenAI-compatible)
openai==1.51.0
httpx==0.27.2

# Redis
redis[hiredis]==5.1.1

# Document parsing
pypdf==4.3.1
python-docx==1.1.2
python-multipart==0.0.12

# Web crawling
playwright==1.47.0
beautifulsoup4==4.12.3

# Text processing
tiktoken==0.7.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
